fix: 모달 컴포넌트 이름/파일 위치 변경 및 잘못된 파라미터 값 수정
수정사항
- ModalRequest.js
  - ConnectRequest.js를 ModalRequest.js로 이름 변경 및 파일 위치 이동
  - 파라미터 기본값 삭제
  - 중복 존재하던 모달 컴포넌트 삭제 (component/connect/ConnectRequest.js 제거)
- ConnectProfileChatRequest.js, HomeCardBack.jsx, ConnectProfilePage.js, FindPasswordPage.js
  - 파일 이름 및 위치 변경에 따른 수정
  - 잘못 전달된 파라미터 값 수정
