+++ src/components/common/ModalRequest.js
@@ -1,28 +1,56 @@
 import React, { useState, useEffect } from "react";
 import { View, Text, StyleSheet } from "react-native";
 import Modal from "react-native-modal";
-import { useTranslation } from "react-i18next";
 
 import { CustomTheme } from "@styles/CustomTheme";
 
 import CompleteIcon from "@components/common/CompleteIcon";
 import ConnectRequestIcon from "@components/common/ConnectRequestIcon";
 
-const ConnectRequest = ({ modalVisible }) => {
-	const { t } = useTranslation();
-
+const ModalRequest = ({
+	modalVisible,
+	setModalVisible,
+	textLoading,
+	textComplete,
+}) => {
 	const [showConnectRequest, setShowConnectRequest] = useState(true);
 	const [showConnectComplete, setShowConnectComplete] = useState(false);
 
 	useEffect(() => {
+		let timeoutRequest, timeoutComplete;
+
 		if (modalVisible) {
 			setShowConnectRequest(true);
 			setShowConnectComplete(false);
+
+			timeoutRequest = setTimeout(() => {
+				setShowConnectRequest(false);
+				setShowConnectComplete(true);
+			}, 2500);
+
+			timeoutComplete = setTimeout(() => {
+				setModalVisible(false);
+			}, 4500);
 		}
+
+		return () => {
+			clearTimeout(timeoutRequest);
+			clearTimeout(timeoutComplete);
+		};
 	}, [modalVisible]);
 
+	const handleBackdropPress = () => {
+		if (!showConnectRequest) {
+			setModalVisible(false);
+		}
+	};
+
 	return (
-		<Modal isVisible={modalVisible} style={styles.modal}>
+		<Modal
+			isVisible={modalVisible}
+			style={styles.modal}
+			onBackdropPress={handleBackdropPress}
+		>
 			<View style={styles.rectangle}>
 				{showConnectRequest && (
 					<View style={styles.connectContainer}>
@@ -31,7 +59,7 @@ const ConnectRequest = ({ modalVisible }) => {
 						</View>
 						<View style={styles.connectTextView}>
 							<Text style={styles.connectText}>
-								{t("connectRequestInProgress")}
+								{textLoading}
 							</Text>
 						</View>
 					</View>
@@ -42,7 +70,7 @@ const ConnectRequest = ({ modalVisible }) => {
 						<CompleteIcon isConnect={true} />
 						<View style={styles.connectTextView}>
 							<Text style={styles.connectText}>
-								{t("connectRequestComplete")}
+								{textComplete}
 							</Text>
 						</View>
 					</View>
@@ -85,4 +113,4 @@ const styles = StyleSheet.create({
 	},
 });
 
-export default ConnectRequest;
+export default ModalRequest;
